Теперь это работает с float и соответствет MISRA, по версии deepseak

+++ 2_convert_celsius_to_farenheit.c
@@ -7,12 +7,12 @@ int main(void)
     const int STEP = 20;
 
     int degrees_fahrenheit = LOWER_LIMIT;
-    int degrees_celsius = 0;
+    float degrees_celsius = 0;
     
     while (degrees_fahrenheit <= UPPER_LIMIT)
     {
-        degrees_celsius = 5 * (degrees_fahrenheit - 32) / 9;
-        printf("%d\t%d\n", degrees_fahrenheit, degrees_celsius);
+        degrees_celsius = 5.0f * (degrees_fahrenheit - 32.0f) / 9.0f;
+        printf("%3d\t%6.1f\n", degrees_fahrenheit, degrees_celsius);
         degrees_fahrenheit += STEP;
     }
     return 0;
